+++ forui/CHANGELOG.md
@@ -20,6 +20,8 @@
 
 * Change FButton's animation to only start on mouse down and up.
 
+* Fix `FLabel` not showing error message if label and description are null.
+
 
 ## 0.6.1